pyproject.toml
[project]
name = "langgraph-daily-news-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "feedparser>=6.0.11",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "lxml>=5.4.0",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
